# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Training Center",
        "description": "Educational Lab",
        "license": {
            "name": "Apache 2.4",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo TC API Server"
        }
    ],
    "paths": {
...
        "/trainer": {
            "get": {
                "tags": [
                    "TrainingCenter"
                ],
                "summary": "Get list of trainers",
                "description": "Returns list of trainers",
                "operationId": "c36bf88dd82b6fd308bf8da3132803bb",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Trainer not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TrainingCenter"
                ],
                "summary": "Create",
                "description": "Returns created trainers data",
                "operationId": "67115e5ab2de498c1c1cda348af0aaed",
                "requestBody": {
                    "description": "Pass user registration credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "firstName",
                                    "lastName",
                                    "phone",
                                    "roleId"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "PassWord12345"
                                    },
                                    "firstName": {
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Smith"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "094415263"
                                    },
                                    "taxIdentityNumber": {
                                        "type": "integer",
                                        "example": 5720125
                                    },
                                    "roleId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/trainer/{id}": {
            "put": {
                "tags": [
                    "TrainingCenter"
                ],
                "summary": "Update existing trainer",
                "description": "Returns updated trainer data",
                "operationId": "1cf31d89626d3334d8e2b649b43ed056",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Trainer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass user registration credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "firstName",
                                    "lastName",
                                    "phone",
                                    "roleId"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user1@mail.com"
                                    },
                                    "firstName": {
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Smith"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "094415263"
                                    },
                                    "taxIdentityNumber": {
                                        "type": "integer",
                                        "example": 5720125
                                    },
                                    "roleId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TrainingCenter"
                ],
                "summary": "Delete existing trainer",
                "description": "Deletes a record",
                "operationId": "fc38ffffdc242b27e36c6b0d0c703134",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Trainer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/period": {
            "get": {
                "tags": [
                    "TrainingCenter"
                ],
                "summary": "Get list of trainers with created data",
                "description": "Returns list of trainers",
                "operationId": "6222111b4a46657eec1960de67e3c70f",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Trainer not found."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}